Split speculative decoding utilities
