+++ mlx_vlm/speculative/dflash.py
@@ -1,0 +1,319 @@
+from typing import Any, Callable, Generator, List, Optional, Tuple
+
+import mlx.core as mx
+import mlx.nn as nn
+
+from .common import (
+    _dflash_block_total,
+    _record_speculative_round,
+    _speculative_walk,
+    _speculative_walk_batch,
+    generation_stream,
+)
+
+
+def _dflash_next_block_size(
+    draft_model: nn.Module,
+    requested_block_total: int,
+    remaining_budget: int,
+) -> int:
+    """Choose the next DFlash verify block size from recent acceptance.
+
+    DFlash checkpoints advertise a trained block size, usually 16. Treat that
+    as the ceiling and back off quickly when deeper positions are mostly
+    rejected. When acceptance is strong at the current depth, grow back toward
+    the configured ceiling.
+    """
+    block_total = min(requested_block_total, remaining_budget)
+    if block_total <= 1:
+        return block_total
+    if getattr(draft_model, "prefer_requested_block_size", False):
+        return block_total
+
+    accept_lens = getattr(draft_model, "accept_lens", None) or []
+    draft_lens = getattr(draft_model, "draft_lens", None) or []
+    recent = [
+        (float(a), int(d))
+        for a, d in zip(accept_lens[-8:], draft_lens[-8:])
+        if int(d) > 0
+    ]
+    if not recent:
+        return block_total
+
+    current = min(block_total, max(2, recent[-1][1] + 1))
+    min_total = min(block_total, 4)
+    drafted = sum(d for _, d in recent)
+    accepted = sum(a for a, _ in recent)
+    accept_rate = accepted / drafted
+    mean_accept = accepted / len(recent)
+
+    if accept_rate < 0.30 or mean_accept < 2.0:
+        if current >= 8:
+            return max(min_total, min(block_total, current // 2))
+        return max(min_total, min(block_total, current - 2))
+
+    if accept_rate < 0.50:
+        return max(min_total, min(block_total, current - 2))
+
+    full_hits = sum(1 for a, d in recent if a >= d)
+    full_hit_rate = full_hits / len(recent)
+    if accept_rate >= 0.85 and full_hit_rate >= 0.75:
+        return min(block_total, current + 2)
+
+    return min(block_total, current)
+
+
+def _dflash_committed_hidden_segments(
+    hidden_full: mx.array, new_tokens_list: List[List[int]]
+) -> List[mx.array]:
+    return [
+        hidden_full[i : i + 1, : len(new_tokens), :]
+        for i, new_tokens in enumerate(new_tokens_list)
+    ]
+
+
+def _dflash_rounds(
+    model: nn.Module,
+    draft_model: nn.Module,
+    prompt_cache: List[Any],
+    hidden: mx.array,
+    *,
+    first_bonus: int,
+    max_tokens: int,
+    sampler: Callable[[mx.array], mx.array],
+    draft_block_size: Optional[int] = None,
+    token_dtype: mx.Dtype = mx.int32,
+) -> Generator[Tuple[int, None], None, None]:
+    """DFlash speculative-decoding **round loop**.
+
+    draft → verify → walk → rollback. ``generate_step`` is responsible
+    for prefill, sampling the first bonus token, and packaging the
+    captured hidden states into ``hidden``.
+    """
+    lm = model.language_model if hasattr(model, "language_model") else model
+    if not hasattr(lm, "rollback_speculative_cache"):
+        raise RuntimeError(
+            f"{type(lm).__name__} does not implement rollback_speculative_cache. "
+            "Speculative decoding with a DFlash drafter currently only "
+            "supports mlx_vlm.models.qwen3_5."
+        )
+
+    target_layer_ids = list(draft_model.config.target_layer_ids)
+    block_total = _dflash_block_total(draft_model, draft_block_size)
+    draft_cache = draft_model.reset(model)
+
+    b = first_bonus
+    emitted = 1  # the first bonus has already been yielded by the caller
+
+    while emitted < max_tokens:
+        bs = _dflash_next_block_size(
+            draft_model,
+            block_total,
+            max_tokens - emitted + 1,
+        )
+        if bs <= 1:
+            break
+
+        draft_tokens = draft_model.draft_block(
+            b, hidden, draft_cache, bs, sampler, token_dtype
+        )
+        mx.async_eval(draft_tokens)
+
+        with mx.stream(generation_stream):
+            verify_input = mx.concatenate(
+                [mx.array([[b]], dtype=token_dtype), draft_tokens],
+                axis=1,
+            )
+            verify_out = lm(
+                verify_input,
+                cache=prompt_cache,
+                capture_layer_ids=target_layer_ids,
+            )
+            hidden = mx.concatenate(verify_out.hidden_states, axis=-1)
+            target_tokens = sampler(verify_out.logits)
+        mx.async_eval(target_tokens, hidden)
+
+        # Walk
+        accepted, new_tokens = _speculative_walk(
+            draft_tokens, target_tokens, max_tokens - emitted
+        )
+        _record_speculative_round(draft_model, accepted, bs - 1)
+
+        # Emit
+        for tok in new_tokens:
+            yield tok, None
+            emitted += 1
+            if emitted >= max_tokens:
+                return
+
+        if accepted < bs - 1:
+            hidden = hidden[:, : accepted + 1, :]
+        b = new_tokens[-1] if new_tokens else b
+
+        if accepted < bs - 1:
+            with mx.stream(generation_stream):
+                lm.rollback_speculative_cache(
+                    prompt_cache, verify_out.gdn_states, accepted, bs
+                )
+
+        if emitted % 256 == 0:
+            mx.clear_cache()
+
+
+def _dflash_rounds_batch(
+    model: nn.Module,
+    draft_model: nn.Module,
+    prompt_cache: List[Any],
+    hidden: mx.array,
+    *,
+    first_bonus: mx.array,
+    max_tokens: int,
+    sampler: Callable[[mx.array], mx.array],
+    draft_block_size: Optional[int] = None,
+    token_dtype: mx.Dtype = mx.int32,
+    stop_check: Optional[Callable[[int, int], bool]] = None,
+) -> Generator[Tuple[List[Optional[int]], None], None, None]:
+    """Batch DFlash speculative-decoding round loop (B > 1).
+
+    Supports continuous batching: when a sequence finishes (EOS or
+    max_tokens), it is filtered out of the target caches and the
+    drafter cache is reinitialized for the new batch size.
+
+    ``stop_check(seq_idx, token_id) -> bool`` is an optional callback
+    that returns True to stop a sequence (e.g. EOS detection).
+
+    Yields ``(tokens_list, None)`` where ``tokens_list[i]`` is the
+    token for sequence ``i`` (or ``None`` if that sequence has nothing
+    to emit this step).
+    """
+    lm = model.language_model if hasattr(model, "language_model") else model
+    if not hasattr(lm, "rollback_speculative_cache"):
+        raise RuntimeError(
+            f"{type(lm).__name__} does not implement " "rollback_speculative_cache."
+        )
+
+    B = first_bonus.shape[0]
+    target_layer_ids = list(draft_model.config.target_layer_ids)
+    block_total = _dflash_block_total(draft_model, draft_block_size)
+    draft_model.reset(model)
+    draft_caches = [draft_model.make_cache() for _ in range(B)]
+
+    # Per-sequence state tracked by ORIGINAL index so the caller sees
+    # stable indices in the yielded token lists.
+    b = first_bonus.tolist()  # active bonus tokens
+    emitted = [1] * B
+    finished = [False] * B
+    active_idx = list(range(B))  # maps active-slot → original-index
+    hidden_by_orig = [hidden[i : i + 1] for i in range(B)]
+
+    total_emitted = sum(emitted)
+
+    while len(active_idx) > 0:
+        remaining = [
+            max(1, max_tokens - emitted[active_idx[j]] + 1)
+            for j in range(len(active_idx))
+        ]
+        bs = _dflash_next_block_size(draft_model, block_total, min(remaining))
+        if bs <= 1:
+            break
+
+        n_active = len(active_idx)
+        b_active = [b[active_idx[j]] for j in range(n_active)]
+        b_arr = mx.array(b_active, dtype=token_dtype)
+
+        # Draft rowwise: the DFlash drafter cache is scalar-offset and has
+        # proven unsafe as a single batched cache on MLX/Metal. Target verify
+        # remains batched below.
+        draft_tokens = mx.concatenate(
+            [
+                draft_model.draft_block(
+                    int(b_active[j]),
+                    hidden_by_orig[active_idx[j]],
+                    draft_caches[active_idx[j]],
+                    bs,
+                    sampler,
+                    token_dtype,
+                )
+                for j in range(n_active)
+            ],
+            axis=0,
+        )
+        mx.async_eval(draft_tokens)
+
+        # Verify
+        with mx.stream(generation_stream):
+            verify_input = mx.concatenate([b_arr[:, None], draft_tokens], axis=1)
+            verify_out = lm(
+                verify_input,
+                cache=prompt_cache,
+                capture_layer_ids=target_layer_ids,
+            )
+            hidden_full = mx.concatenate(verify_out.hidden_states, axis=-1)
+            target_tokens = sampler(verify_out.logits)
+        mx.async_eval(target_tokens, hidden_full)
+
+        # Walk (per-sequence)
+        budgets = [max_tokens - emitted[active_idx[j]] for j in range(n_active)]
+        accepted_list, new_tokens_list = _speculative_walk_batch(
+            draft_tokens, target_tokens, budgets
+        )
+
+        min_accepted = min(accepted_list)
+        accepted_arr = mx.array(accepted_list)
+
+        hidden_segments = _dflash_committed_hidden_segments(
+            hidden_full, new_tokens_list
+        )
+        for j in range(n_active):
+            orig = active_idx[j]
+            if hidden_segments[j].shape[1] > 0:
+                hidden_by_orig[orig] = hidden_segments[j]
+
+        for a in accepted_list:
+            _record_speculative_round(draft_model, a, bs - 1)
+
+        # Emit (map active slots back to original indices)
+        max_new = max(len(nt) for nt in new_tokens_list) if new_tokens_list else 0
+        for pos in range(max_new):
+            tokens_out: List[Optional[int]] = [None] * B
+            for j in range(n_active):
+                orig = active_idx[j]
+                if pos < len(new_tokens_list[j]) and not finished[orig]:
+                    tok = new_tokens_list[j][pos]
+                    tokens_out[orig] = tok
+                    emitted[orig] += 1
+                    if emitted[orig] >= max_tokens:
+                        finished[orig] = True
+                    if stop_check is not None and stop_check(orig, tok):
+                        finished[orig] = True
+            yield tokens_out, None
+
+        # Update bonus tokens
+        for j in range(n_active):
+            orig = active_idx[j]
+            if new_tokens_list[j]:
+                b[orig] = new_tokens_list[j][-1]
+
+        if min_accepted < bs - 1:
+            with mx.stream(generation_stream):
+                lm.rollback_speculative_cache(
+                    prompt_cache, verify_out.gdn_states, accepted_arr, bs
+                )
+
+        # --- Continuous batching: filter out finished sequences ---
+        keep_slots = [j for j in range(n_active) if not finished[active_idx[j]]]
+        if len(keep_slots) < n_active:
+            if len(keep_slots) == 0:
+                break
+            # Filter target caches (BatchKVCache supports this)
+            keep_mx = mx.array(keep_slots, dtype=mx.int32)
+            for c in prompt_cache:
+                if hasattr(c, "filter"):
+                    c.filter(keep_mx)
+            # Update active index mapping
+            active_idx = [active_idx[j] for j in keep_slots]
+
+        new_total = sum(emitted)
+        if new_total // 256 > total_emitted // 256:
+            mx.clear_cache()
+        total_emitted = new_total
